project/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.14
requests==2.31.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.0